HMM & the Viterbi Algorithm: Python Implementation
build an HMM POS tagger using the Penn Treebank dataset as the training corpus.

Conduct exploratory analysis on a tagged corpus
Sample the data into 70:30 train-test sets
Train an HMM model using the tagged corpus:
Calculating the emission probabilities 
Calculating the transition probabilities 
Write the Viterbi algorithm to POS tag sequence of words (sentence)
Evaluate the model predictions against the ground truth 


Result:
The Viterbi algorithm gave us approximately 87% accuracy. The 13% loss of accuracy, to a large part, can be attributed to the fact that when the algorithm hits an unknown word (i.e. not present in the training set), it naively assigns the first tag in the list of tags that we have created. In this case, the first tag happens to be JJ, so it assigns JJ to all unknown words.