training CNNs on images where an "effusion" is present in XRAYS.
Building an end-to-end pipeline for training CNNs on images where an "effusion" is present. This is a 		classification problem, where we will be dealing with two classes - 'effusion' and ‘nofinding’. The 		ratio of 'effusion' vs 'nofinding' is almost 10 (107/1000) hense the data is highly imbalanced.Hense 		the evaluation matrices AUC(Area under curve)/'weighted categorical cross-entropy'. This is a 			measure of loss, which applies weights to different forms of errors is used.

                AUC in normal classification. : 0.68
                AUC on weighted categorical cross-entropy: 0.73436